Refactor shared system test helpers and harden values access
