+++ spec/system/haya-select-pagination-spec.js
@@ -2,181 +2,19 @@ import "velocious/build/src/testing/test.js"
 import timeout from "awaitery/build/timeout.js"
 import waitFor from "awaitery/build/wait-for.js"
 import SystemTest from "system-testing/build/system-test.js"
+import {
+  clickPaginationSelector,
+  closePaginatedSelect,
+  findPaginationPageButton,
+  openPaginatedSelect,
+  setPaginationInputValue,
+  waitForPaginationLabel
+} from "./pagination-system-test-helpers.js"
+import {setupSystemTestLifecycle, systemTestArgs} from "./system-test-lifecycle.js"
 
 import HayaSelectSystemTestHelper from "../../src/system-test-helpers.js"
 
-SystemTest.rootPath = "/?systemTest=true"
-const systemTestHttpHost = process.env.SYSTEM_TEST_HTTP_HOST || "127.0.0.1"
-const systemTestHttpConnectHost = process.env.SYSTEM_TEST_HTTP_CONNECT_HOST || systemTestHttpHost
-const systemTestArgs = {
-  debug: true,
-  httpConnectHost: systemTestHttpConnectHost,
-  httpHost: systemTestHttpHost
-}
-let didStartSystemTest = false
-
-beforeAll(async () => {
-  const systemTest = SystemTest.current(systemTestArgs)
-  if (!systemTest.isStarted()) {
-    await timeout({errorMessage: "beforeAll: timed out starting SystemTest", timeout: 30000}, async () => {
-      await systemTest.start()
-    })
-    didStartSystemTest = true
-  }
-  systemTest.setBaseSelector("[data-testid='systemTestingComponent']")
-})
-
-afterAll(async () => {
-  if (!didStartSystemTest) return
-
-  await timeout({errorMessage: "afterAll: timed out stopping SystemTest", timeout: 30000}, async () => {
-    await SystemTest.current().stop()
-  })
-})
-
-const setPaginationInputValue = async (systemTest, value) => {
-  await waitFor({timeout: 5000}, async () => {
-    const element = await systemTest.find("[data-class='pagination-input']", {useBaseSelector: false})
-    const driver = systemTest.getDriver()
-    await driver.executeScript(
-      "arguments[0].focus(); arguments[0].value = arguments[1]; arguments[0].dispatchEvent(new Event('input', {bubbles: true})); arguments[0].dispatchEvent(new Event('change', {bubbles: true})); arguments[0].blur();",
-      element,
-      String(value)
-    )
-  })
-}
-
-const openPaginatedSelect = async (systemTest) => {
-  const selectContainer = await systemTest.find("[data-testid='hayaSelectPaginationRoot'] [data-class='select-container']")
-  const driver = systemTest.getDriver()
-  const searchInputs = await systemTest.all(
-    "[data-testid='hayaSelectPaginationRoot'] [data-class='search-text-input']",
-    {timeout: 0, visible: true}
-  )
-
-  if (searchInputs.length > 0) {
-    await driver.executeScript(
-      "arguments[0].scrollIntoView({block: 'center', inline: 'center'})",
-      selectContainer
-    )
-    await systemTest.click(selectContainer)
-    await waitFor({timeout: 5000}, async () => {
-      const searchInputsAfterClick = await systemTest.all(
-        "[data-testid='hayaSelectPaginationRoot'] [data-class='search-text-input']",
-        {timeout: 0, visible: true}
-      )
-
-      if (searchInputsAfterClick.length > 0) {
-        throw new Error("Search input still visible")
-      }
-    })
-  }
-
-  await driver.executeScript(
-    "arguments[0].scrollIntoView({block: 'center', inline: 'center'})",
-    selectContainer
-  )
-  await systemTest.click(selectContainer)
-  await waitFor({timeout: 5000}, async () => {
-    const searchInputsAfterClick = await systemTest.all(
-      "[data-testid='hayaSelectPaginationRoot'] [data-class='search-text-input']",
-      {timeout: 0, visible: true}
-    )
-
-    if (searchInputsAfterClick.length === 0) {
-      throw new Error("Search input not visible yet")
-    }
-  })
-
-  await waitFor({timeout: 5000}, async () => {
-    const paginationElements = await systemTest.all("[data-class='options-pagination']", {timeout: 0, useBaseSelector: false})
-
-    if (paginationElements.length === 0) {
-      throw new Error("Pagination not visible yet")
-    }
-  })
-
-  await paginationLabelText(systemTest)
-}
-
-const closePaginatedSelect = async (systemTest) => {
-  const inputs = await systemTest.all(
-    "[data-testid='hayaSelectPaginationRoot'] [data-class='search-text-input']",
-    {timeout: 0, visible: true}
-  )
-
-  if (inputs.length > 0) {
-    const selectContainer = await systemTest.find("[data-testid='hayaSelectPaginationRoot'] [data-class='select-container']")
-    const driver = systemTest.getDriver()
-    await driver.executeScript(
-      "arguments[0].scrollIntoView({block: 'center', inline: 'center'})",
-      selectContainer
-    )
-    await systemTest.click(selectContainer)
-  }
-}
-
-const paginationLabelText = async (systemTest) => {
-  const labelElements = await systemTest.all("[data-class='pagination-input']", {timeout: 0, useBaseSelector: false})
-
-  if (labelElements.length === 0) return null
-
-  const value = await labelElements[0].getAttribute("value")
-
-  return value ? value.trim() : ""
-}
-
-const waitForPaginationLabel = async (systemTest, expectedText) => {
-  const expectedPage = Number(expectedText.match(/Page (\d+) of/)?.[1])
-  await waitFor({timeout: 5000}, async () => {
-    const labelText = await paginationLabelText(systemTest)
-
-    if (labelText === expectedText) return
-
-    if (Number.isFinite(expectedPage) && labelText === String(expectedPage)) return
-    if (Number.isFinite(expectedPage) && Number(labelText) === expectedPage) return
-
-    if (!Number.isFinite(expectedPage) && labelText === expectedText) return
-
-    if (labelText !== expectedText) {
-      throw new Error(`Unexpected pagination label: ${labelText}`)
-    }
-  })
-}
-
-const findPaginationPageButton = async (systemTest, pageNumber) => {
-  let matchingButton
-
-  await waitFor({timeout: 5000}, async () => {
-    const pageButtons = await systemTest.all("[data-class='pagination-page']", {useBaseSelector: false})
-
-    for (const pageButton of pageButtons) {
-      const buttonText = (await pageButton.getText()).trim()
-
-      if (buttonText === String(pageNumber)) {
-        matchingButton = pageButton
-        return
-      }
-    }
-
-    throw new Error(`Pagination button not found for page ${pageNumber}`)
-  })
-
-  return matchingButton
-}
-
-const clickPaginationSelector = async (systemTest, selector) => {
-  await waitFor({timeout: 5000}, async () => {
-    const element = await systemTest.find(selector, {useBaseSelector: false})
-    const driver = systemTest.getDriver()
-    await driver.executeScript(
-      "arguments[0].scrollIntoView({block: 'center', inline: 'center'})",
-      element
-    )
-    await driver.executeScript("arguments[0].focus()", element)
-    await systemTest.click(element)
-  })
-}
+setupSystemTestLifecycle()
 
 describe("HayaSelect pagination", () => {
   afterEach(async () => {
