+++ spec/system/haya-select-render-spec.js
@@ -2,37 +2,11 @@ import "velocious/build/src/testing/test.js"
 import timeout from "awaitery/build/timeout.js"
 import waitFor from "awaitery/build/wait-for.js"
 import SystemTest from "system-testing/build/system-test.js"
+import {setupSystemTestLifecycle, systemTestArgs} from "./system-test-lifecycle.js"
 
 import HayaSelectSystemTestHelper from "../../src/system-test-helpers.js"
 
-SystemTest.rootPath = "/?systemTest=true"
-const systemTestHttpHost = process.env.SYSTEM_TEST_HTTP_HOST || "127.0.0.1"
-const systemTestHttpConnectHost = process.env.SYSTEM_TEST_HTTP_CONNECT_HOST || systemTestHttpHost
-const systemTestArgs = {
-  debug: true,
-  httpConnectHost: systemTestHttpConnectHost,
-  httpHost: systemTestHttpHost
-}
-let didStartSystemTest = false
-
-beforeAll(async () => {
-  const systemTest = SystemTest.current(systemTestArgs)
-  if (!systemTest.isStarted()) {
-    await timeout({errorMessage: "beforeAll: timed out starting SystemTest", timeout: 30000}, async () => {
-      await systemTest.start()
-    })
-    didStartSystemTest = true
-  }
-  systemTest.setBaseSelector("[data-testid='systemTestingComponent']")
-})
-
-afterAll(async () => {
-  if (!didStartSystemTest) return
-
-  await timeout({errorMessage: "afterAll: timed out stopping SystemTest", timeout: 30000}, async () => {
-    await SystemTest.current().stop()
-  })
-})
+setupSystemTestLifecycle()
 
 describe("HayaSelect", () => {
   it("renders in the example app", async () => {
